requirements_docker.txt
# Django 5.2 - Latest stable version
Django>=5.2,<5.3

# Image processing
Pillow>=10.0.0

# PDF processing for book optimization
pypdf>=4.0.0
pdf2image>=1.17.0